Add GHC/ExecutionStack.hsc
This is a small API for examining the Execution Stack, including getting
stack frames with source-level information using the DWARF debug data.
For ghc to generate the debug data in the first place, the '-g' flag
should be passed to ghc when compiling your modules.

It took me a few iterations to create this module, yet I decided to
"erase history" and make this into one commit. My failed attempts can be
found here:
https://github.com/Tarrasch/packages-base/commits/gh-lifeline

Test cases will follow in a subsequent commit.